document-iq/document-iq-platform/components/account-component/src/document_iq_platform_account/security/jwt.py
from datetime import datetime, timedelta
from jose import jwt
from platform_shared.config.settings import Settings

settings = Settings()


def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(
        minutes=settings.access_token_expire_minutes
    )
    to_encode.update({"exp": expire})

    return jwt.encode(
        to_encode,
        settings.jwt_secret,
        algorithm=settings.jwt_algorithm,
    )
