document-iq/document-iq-platform/components/account-component/pyproject.toml
[project]
name = "document-iq-platform-account"
version = "1.0.1"
description = "Document aggregator service for DocumentIQ platform"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "kafka-python",
    "pydantic",
    "python-dotenv",
    "redis",
    "sqlalchemy",
    "alembic",
    "passlib",
    "python-jose",
    "pymysql",
    "document-iq-platform-shared"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]